Phase2/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
python-dotenv==1.0.0
bcrypt==4.0.1
argon2-cffi==23.1.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
Pillow==10.0.0
requests==2.31.0
drf-yasg==1.21.7